+++ src/bson.ts
@@ -54,7 +54,8 @@ export {
   MaxKey,
   BSONRegExp,
   Decimal128,
-  NumberUtils
+  NumberUtils,
+  ByteUtils
 };
 export { BSONValue, bsonType, type BSONTypeTag } from './bson_value';
 export { BSONError, BSONVersionError, BSONRuntimeError, BSONOffsetError } from './error';