+++ src/utils/byte_utils.ts
@@ -9,18 +9,26 @@ import { webByteUtils } from './web_byte_utils';
  * ByteUtils is configured at load time to use Node.js or Web based APIs for the internal implementations.
  */
 export type ByteUtils = {
+  /** Checks if the given value is a Uint8Array. */
+  isUint8Array: (value: unknown) => value is Uint8Array;
   /** Transforms the input to an instance of Buffer if running on node, otherwise Uint8Array */
   toLocalBufferType: (buffer: Uint8Array | ArrayBufferView | ArrayBuffer) => Uint8Array;
   /** Create empty space of size */
   allocate: (size: number) => Uint8Array;
   /** Create empty space of size, use pooled memory when available */
   allocateUnsafe: (size: number) => Uint8Array;
+  /** Compare 2 Uint8Arrays lexicographically */
+  compare: (buffer1: Uint8Array, buffer2: Uint8Array) => -1 | 0 | 1;
+  /** Concatenating all the Uint8Arrays in new Uint8Array. */
+  concat: (list: Uint8Array[]) => Uint8Array;
   /** Check if two Uint8Arrays are deep equal */
   equals: (a: Uint8Array, b: Uint8Array) => boolean;
-  /** Check if two Uint8Arrays are deep equal */
+  /** Create a Uint8Array from an array of numbers */
   fromNumberArray: (array: number[]) => Uint8Array;
   /** Create a Uint8Array from a base64 string */
   fromBase64: (base64: string) => Uint8Array;
+  /** Create a Uint8Array from a UTF8 string */
+  fromUTF8: (utf8: string) => Uint8Array;
   /** Create a base64 string from bytes */
   toBase64: (buffer: Uint8Array) => string;
   /** **Legacy** binary strings are an outdated method of data transfer. Do not add public API support for interpreting this format */
@@ -58,6 +66,7 @@ const hasGlobalBuffer = typeof Buffer === 'function' && Buffer.prototype?._isBuf
  * The type annotation is important here, it asserts that each of the platform specific
  * utils implementations are compatible with the common one.
  *
- * @internal
+ * @public
+ * @experimental
  */
 export const ByteUtils: ByteUtils = hasGlobalBuffer ? nodeJsByteUtils : webByteUtils;